feat(site): migrate layout and home page to new site structure
