--- src/app/(home)/layout.tsx
@@ -1,19 +1,0 @@
-import type { ReactNode } from 'react';
-import { HomeLayout } from 'fumadocs-ui/layouts/home';
-import { baseOptions } from '@/app/layout.config';
-
-export default function Layout({ children }: { children: ReactNode }) {
-  return (
-    <HomeLayout
-      {...baseOptions}
-      links={[
-        {
-          text: 'Documentation',
-          url: '/docs',
-        },
-      ]}
-    >
-      {children}
-    </HomeLayout>
-  );
-}
